recommendation_system/requirements.txt
# FastAPI и сервер
fastapi
uvicorn[standard]

# Модель и математика
scikit-learn
scipy
numpy
pandas

# База данных
SQLAlchemy
psycopg2-binary

# Kafka
kafka-python
aiokafka

# Планировщик
APScheduler

# Вспомогательные
python-dotenv
requests
tqdm

asyncpg
